requirements.txt
Flask>=2.3
Werkzeug>=2.2,<3
Flask-SQLAlchemy>=3.0
Flask-Migrate==3.1.0
Flask-JWT-Extended==4.4.0
pandas==1.4.2
# reportlab==3.6.12
python-docx==0.8.11
pytest==7.1.2
pytest-flask==1.2.0
marshmallow==3.14.1
